pull extractValueFromCustomField to its own file
