+++ src/converters/IssueEntityConverter.ts
@@ -1,5 +1,5 @@
+import { IntegrationLogger } from "@jupiterone/jupiter-managed-integration-sdk";
 import camelCase from "lodash/camelCase";
-import { isArray } from "util";
 
 import {
   CHANGE_ISSUE_ENTITY_CLASS,
@@ -11,10 +11,11 @@ import {
   RISK_ISSUE_ENTITY_CLASS,
   VULN_ISSUE_ENTITY_CLASS,
 } from "../entities";
-import { Field, Issue, TextContent } from "../jira";
+import { Field, Issue } from "../jira";
 import parseContent from "../jira/parseContent";
 import generateEntityKey from "../utils/generateEntityKey";
 import getTime from "../utils/getTime";
+import { extractValueFromCustomField } from "./extractValueFromCustomField";
 
 const DONE = [
   "done",
@@ -30,93 +31,9 @@ const DONE = [
   "transferred",
 ];
 
-function parseNumber(s: string | number): number | string {
-  if (typeof s !== "string") {
-    return s;
-  }
-  const NUM_REGEX = /^[\d,]*(\.[\d]*)?(e[\d]*)?$/;
-  const match = s.match(NUM_REGEX);
-  if (match) {
-    const numStr = s.replace(",", "");
-    return match[1] || match[2] ? parseFloat(numStr) : parseInt(numStr, 10);
-  } else {
-    return s;
-  }
-}
-
-/**
- * There are several different custom field types Jira allows.
- * Each one stores information in a differnt way. This makes sure
- * we extract the most relevant data for each type.
- *
- * https://developer.atlassian.com/server/jira/platform/jira-rest-api-examples/#setting-custom-field-data-for-other-field-types
- */
-export function extractValueFromCustomField(value: any): any {
-  /**
-   * Handles:
-   *   DatePickerFields
-   *   DateTimeFields
-   *   FreeTextFields
-   *   NumberFields
-   *   TextFields
-   *   URLFields
-   */
-  if (["string", "number"].includes(typeof value)) {
-    return value;
-  } else if (typeof value === "object") {
-    /**
-     * I'm unsure what this handles as this case is not in the Jiras docs,
-     * but I'm too scared to pull it out.
-     */
-    if (value.type === "doc" && value.content) {
-      return parseContent(value.content as TextContent[]);
-    } else if (value.value) {
-      /**
-       * Handles:
-       *   CascadingSelectFields (minus the child values)
-       *   RadioButtons
-       *   SelectLists
-       */
-      return typeof value.value === "object"
-        ? JSON.stringify(value.value)
-        : parseNumber(value.value);
-    } else if (value.name || value.id || value.key) {
-      /**
-       * Unsuppoted Custom Fields:
-       *   ProjectPickers
-       *   SingleVersionPickers
-       *   UserPickers
-       *   VersionPickers
-       */
-      if (value.name) {
-        return value.name;
-      }
-      if (value.id) {
-        return value.id;
-      }
-      if (value.key) {
-        return value.key;
-      }
-    } else if (isArray(value)) {
-      /**
-       * Handles:
-       *   Multi Group Picker
-       *   Multi Select
-       *   Multi User Picker
-       */
-      return value.map(extractValueFromCustomField).join(",");
-    } else {
-      // We may want to throw an error here
-      return undefined;
-    }
-  } else {
-    // We may want to throw an error here
-    return undefined;
-  }
-}
-
 export function createIssueEntity(
   issue: Issue,
+  logger: IntegrationLogger,
   fieldsById: { [id: string]: Field } = {},
   customFieldsToInclude: string[] = [],
 ): IssueEntity {
@@ -131,7 +48,7 @@ export function createIssueEntity(
         customFieldsToInclude.includes(key) ||
         customFieldsToInclude.includes(fieldName)
       ) {
-        customFields[fieldName] = extractValueFromCustomField(value);
+        customFields[fieldName] = extractValueFromCustomField(value, logger);
       }
     }
   }
