+++ src/converters/extractValueFromCustomField.test.ts
@@ -1,0 +1,103 @@
+import { createTestLogger } from "@jupiterone/jupiter-managed-integration-sdk";
+import { extractValueFromCustomField } from "./extractValueFromCustomField";
+
+describe("extractValueFromCustomField", () => {
+  // All custom Jira field types
+  // https://developer.atlassian.com/server/jira/platform/jira-rest-api-examples/#setting-custom-field-data-for-other-field-types
+  const CascadingSelectField = { value: "green", child: { value: "blue" } };
+  const DatePickerField = "2011-10-03";
+  const DateTimeField = "2011-10-19T10:29:29.908+1100";
+  const FreeTextField = "Free text goes here.  Type away!";
+  const GroupPicker = { name: "jira-developers" };
+  const MultiGroupPicker = [
+    { name: "admins" },
+    { name: "jira-developers" },
+    { name: "jira-users" },
+  ];
+  const MultiSelect = [{ value: "red" }, { value: "blue" }, { value: "green" }];
+  const MultiUserPicker = [
+    { name: "charlie" },
+    { name: "bjones" },
+    { name: "tdurden" },
+  ];
+  const NumberField = 42.07;
+  const ProjectPicker = { key: "JRADEV" };
+  const RadioButtons = { value: "red" };
+  const SelectList = { value: "red" };
+  const SingleVersionPicker = { name: "5.0" };
+  const TextField = "Is anything better than text?";
+  const URLField = "http://www.atlassian.com";
+  const UserPicker = { name: "brollins" };
+  const VersionPicker = [{ name: "1.0" }, { name: "1.1.1" }, { name: "2.0" }];
+
+  const logger = createTestLogger();
+
+  it("should extact the value of a CascadingSelectField correctly", () => {
+    const value = extractValueFromCustomField(CascadingSelectField, logger);
+    expect(value).toEqual("green");
+  });
+  it("should extact the value of a DatePickerField correctly", () => {
+    const value = extractValueFromCustomField(DatePickerField, logger);
+    expect(value).toEqual("2011-10-03");
+  });
+  it("should extact the value of a DateTimeField correctly", () => {
+    const value = extractValueFromCustomField(DateTimeField, logger);
+    expect(value).toEqual("2011-10-19T10:29:29.908+1100");
+  });
+  it("should extact the value of a FreeTextField correctly", () => {
+    const value = extractValueFromCustomField(FreeTextField, logger);
+    expect(value).toEqual("Free text goes here.  Type away!");
+  });
+  it("should extact the value of a GroupPicker correctly", () => {
+    const value = extractValueFromCustomField(GroupPicker, logger);
+    expect(value).toEqual("jira-developers");
+  });
+  it("should extact the value of a MultiGroupPicker correctly", () => {
+    const value = extractValueFromCustomField(MultiGroupPicker, logger);
+    expect(value).toEqual("admins,jira-developers,jira-users");
+  });
+  it("should extact the value of a MultiSelect correctly", () => {
+    const value = extractValueFromCustomField(MultiSelect, logger);
+    expect(value).toEqual("red,blue,green");
+  });
+  it("should extact the value of a MultiUserPicker correctly", () => {
+    const value = extractValueFromCustomField(MultiUserPicker, logger);
+    expect(value).toEqual("charlie,bjones,tdurden");
+  });
+  it("should extact the value of a NumberField correctly", () => {
+    const value = extractValueFromCustomField(NumberField, logger);
+    expect(value).toEqual(42.07);
+  });
+  it("should extact the value of a ProjectPicker correctly", () => {
+    const value = extractValueFromCustomField(ProjectPicker, logger);
+    expect(value).toEqual("JRADEV");
+  });
+  it("should extact the value of a RadioButtons correctly", () => {
+    const value = extractValueFromCustomField(RadioButtons, logger);
+    expect(value).toEqual("red");
+  });
+  it("should extact the value of a SelectList correctly", () => {
+    const value = extractValueFromCustomField(SelectList, logger);
+    expect(value).toEqual("red");
+  });
+  it("should extact the value of a SingleVersionPicker correctly", () => {
+    const value = extractValueFromCustomField(SingleVersionPicker, logger);
+    expect(value).toEqual("5.0");
+  });
+  it("should extact the value of a TextField correctly", () => {
+    const value = extractValueFromCustomField(TextField, logger);
+    expect(value).toEqual("Is anything better than text?");
+  });
+  it("should extact the value of a URLField correctly", () => {
+    const value = extractValueFromCustomField(URLField, logger);
+    expect(value).toEqual("http://www.atlassian.com");
+  });
+  it("should extact the value of a UserPicker correctly", () => {
+    const value = extractValueFromCustomField(UserPicker, logger);
+    expect(value).toEqual("brollins");
+  });
+  it("should extact the value of a VersionPicker correctly", () => {
+    const value = extractValueFromCustomField(VersionPicker, logger);
+    expect(value).toEqual("1.0,1.1.1,2.0");
+  });
+});
